requirements.txt
click==8.1.7
prettytable==3.9.0
python-dotenv==1.0.0
pytest==7.4.2